SERVER/requirements.txt
Django>=3.2.4
djangorestframework>=3.12.4
djongo>=1.3.6
django-cors-headers>=3.7.0
pandas>=0.18.1
sqlparse==0.2.4